+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 5.20.0 - (April 15, 2024)
+
 * Changed
   * Updated subsection logic to properly support grid implementations.